fix: 체형 스마트 분석 API 요청 내 잚못된 필드 타입 수정

+++ src/main/java/fittering/mall/controller/dto/request/RequestSmartMeasurementDto.java
@@ -12,9 +12,9 @@ public class RequestSmartMeasurementDto {
     @NonNull
     private String side;
     @NonNull
-    private Integer height;
+    private Double height;
     @NonNull
-    private Integer weight;
+    private Double weight;
     @NonNull
     private String sex;
 }
